pyproject.toml
[project]
name = "femsahacktheshelf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
    "pymongo>=4.13.0",
    "streamlit>=1.45.1",
    "ultralytics>=8.3.138",
]
